+++ CHANGES.rst
@@ -25,6 +25,7 @@ AudioLazy changes history
 
   - Pitch follower via zero-crossing rate with Tkinter GUI
   - Pi with Madhava-Gregory-Leibniz series and Machin formula using Stream
+  - LPC plot with DFT, showing two formants (magnitude peaks)
 
 + general: